PAJ 21/08/2017 - Creada la tabla de relación entre usuarios. - Sólo se puede enviar una petición de amistad. - Mejoras gráficas varias.

+++ src/main/java/dao/UserDAOImpl.java
@@ -5,8 +5,10 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 import model.User;
 
+import java.sql.Date;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
+import java.sql.Types;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -27,9 +29,21 @@ public class UserDAOImpl implements UserDAO {
     }
 
     @Override
+    public User getUserByUserID(int user_id_pk) {
+
+        String query = String.format("SELECT * FROM users WHERE users.user_id_pk = %d", user_id_pk);
+        List<User> userList = getUsers(query);
+
+        if (!(userList.isEmpty()))
+            return userList.get(0);
+
+        return new User();
+    }
+
+    @Override
     public User getUserByUsername(String username) {
 
-        String query = "SELECT * FROM users WHERE users.username = '" + username + "'";
+        String query = String.format("SELECT * FROM users WHERE users.username = '%s'", username);
         List<User> userList = getUsers(query);
 
         if (!(userList.isEmpty()))
@@ -44,12 +58,22 @@ public class UserDAOImpl implements UserDAO {
 
         try {
 
-            String query = "INSERT INTO users (username, password, admin) VALUES(?,?,?)";
+            String query = "INSERT INTO users (username, password, admin, name, surname, gender, country, email, birthday, biography, registration) VALUES(?,?,?,?,?,?,?,?,?,?,?)";
             PreparedStatement ps = db.getConnection(query);
 
             ps.setString(1, user.getUsername());
             ps.setString(2, user.getPassword());
             ps.setBoolean(3, user.isAdmin());
+            ps.setString(4, user.getName());
+            ps.setString(5, user.getSurname());
+            ps.setString(6, user.getGender());
+            ps.setString(7, user.getCountry());
+            ps.setString(8, user.getEmail());
+            if (user.getBirthday() != null) ps.setDate(9, new Date(user.getBirthday().getTime()));
+            else ps.setNull(9, Types.DATE);
+            ps.setString(10, user.getBiography());
+            ps.setDate(11, new Date(user.getRegistration().getTime()));
+
             ps.execute();
             ps.close();
 
@@ -62,11 +86,15 @@ public class UserDAOImpl implements UserDAO {
     public void deleteUser(int user_id_pk) {
 
         try {
+            RelationshipDAO relationshipDAO = new RelationshipDAOImpl();
+            String query = String.format("DELETE FROM users WHERE users.user_id_pk = %d", user_id_pk);
 
-            String query = "DELETE FROM users WHERE users.user_id_pk = ?";
+            // Borramos las posibles relaciones con otros usuarios
+            relationshipDAO.deleteRelationshipsByUserID(user_id_pk);
+
+            // Borramos al usuario
             PreparedStatement ps = db.getConnection(query);
 
-            ps.setInt(1, user_id_pk);
             ps.execute();
             ps.close();
 
@@ -86,7 +114,7 @@ public class UserDAOImpl implements UserDAO {
             ResultSet rs = ps.executeQuery();
 
             while (rs.next()) {
-                userList.add(new User(rs.getInt("user_id_pk"), rs.getString("username"), rs.getString("password"), rs.getBoolean("admin")));
+                userList.add(new User(rs.getInt("user_id_pk"), rs.getString("username"), rs.getString("password"), rs.getBoolean("admin"), rs.getString("name"), rs.getString("surname"), rs.getString("gender"), rs.getString("country"), rs.getString("email"), rs.getDate("birthday"), rs.getString("biography"), rs.getDate("registration")));
             }
 
             ps.close();
